youtube-crawling/requirements.txt
selenium~=4.20.0
webdriver-manager
tqdm~=4.66.4
openai
requests
beautifulsoup4
mysql-connector-python